+++ joljak-application/src/main/java/kr/joljak/api/notice/controller/NoticeController.java
@@ -2,11 +2,17 @@ package kr.joljak.api.notice.controller;
 
 import io.swagger.annotations.ApiOperation;
 import java.util.List;
+import java.util.stream.Collectors;
 import javax.validation.Valid;
 import kr.joljak.api.notice.request.NoticeRequest;
 import kr.joljak.api.notice.response.NoticeResponse;
 import kr.joljak.api.notice.response.NoticesResponse;
-import kr.joljak.api.notice.service.NoticeService;
+import kr.joljak.domain.notice.dto.SimpleNotice;
+import kr.joljak.domain.notice.entity.Notice;
+import kr.joljak.domain.notice.service.NoticeService;
+import kr.joljak.domain.user.entity.User;
+import kr.joljak.domain.user.service.UserService;
+import kr.joljak.domain.util.FetchPages;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -26,12 +32,15 @@ import org.springframework.web.bind.annotation.RestController;
 public class NoticeController {
 
   private final NoticeService noticeService;
+  private final UserService userService;
 
   @ApiOperation("공지사항 생성 API")
   @PostMapping
   @ResponseStatus(HttpStatus.CREATED)
   public NoticeResponse create(@Valid @RequestBody NoticeRequest noticeRequest) {
-    return noticeService.addNotice(noticeRequest);
+    User user = userService.getUserByClassOf(noticeRequest.getClassOf());
+    Notice notice = noticeService.addNotice(noticeRequest.toNotice(user));
+    return getNoticeResponse(SimpleNotice.of(notice));
   }
 
   @ApiOperation("공지사항 조회 API")
@@ -39,14 +48,38 @@ public class NoticeController {
   @ResponseStatus(HttpStatus.OK)
   public NoticesResponse getNotices(@RequestParam int page,
     @RequestParam(defaultValue = "10") int size) {
-    return noticeService.getNoticesByPage(page, size);
+    List<Notice> noticeList = noticeService.getNoticesByPage(page, size);
+    List<NoticeResponse> noticeResponseList = getNoticeResponseListFrom(noticeList);
+    return NoticesResponse.builder()
+      .noticeResponseList(noticeResponseList)
+      .page(FetchPages.of(page, size))
+      .build();
   }
 
   @ApiOperation("공지사항 개별 조회 API")
   @GetMapping("/{id}")
   @ResponseStatus(HttpStatus.OK)
   public NoticeResponse getNotice(@PathVariable("id") Long id) {
-    return noticeService.getNoticeById(id);
+
+    Notice notice = noticeService.getNoticeById(id);
+    return getNoticeResponse(SimpleNotice.of(notice));
+  }
+
+  private List<NoticeResponse> getNoticeResponseListFrom(List<Notice> noticeList) {
+    return noticeList.stream().map(notice -> getNoticeResponse(SimpleNotice.of(notice))).collect(
+      Collectors.toList());
+  }
+
+  private NoticeResponse getNoticeResponse(SimpleNotice simpleNotice) {
+    return NoticeResponse.builder()
+      .id(simpleNotice.getId())
+      .classOf(simpleNotice.getClassOf())
+      .author(simpleNotice.getAuthor())
+      .title(simpleNotice.getTitle())
+      .content(simpleNotice.getContent())
+      .createdDate(simpleNotice.getCreatedDate())
+      .modifiedDate(simpleNotice.getModifiedDate())
+      .build();
   }
 
 }
